Default to using one thread per CPU 🚀
To increase performance with the default settings, default to using one
thread per CPU. The user must explicitly opt out if they want
single-threaded behavior.

This indirectly solves a different issue: the only way to automatically
peg the number of threads to the number of CPU's was to set
`--worker 0 --threads 0` even though the `worker` argument had been
deprecated.

Fixes #18